completed new cache object interface update in application - 1st pass. now debugging.

+++ data/cache/cache_defs.h
@@ -49,8 +49,8 @@ namespace nimbus {
 namespace cache {
 enum CacheAccess { SHARED, EXCLUSIVE };
 enum CacheTType { VAR, STRUCT };
-typedef size_t type_id_t;
-typedef size_t co_id_t;
+typedef uint64_t type_id_t;
+typedef uint64_t co_id_t;
 typedef uint64_t distance_t;
 }  // namespace cache
 }  // namespace nimbus
